rekrutacje/Back-end/todo_api/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
pydantic==1.10.8
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.0.0 